Added SyncVolume to MediaPlayer message enum
Its for use in the subscription loop to better handle the intention of
the subscription. It simply syncs the module's state with what is
reported by playerctl, and does not try to set that volume after. This
avoids noisy logs of when the current player can't have its volume set
through playerctl.