Add clock format cycling feature
Add support for multiple datetime formats in the clock module that can be cycled through by clicking. When the `formats` array is configured, clicking the clock cycles through each format in sequence. The update interval automatically adjusts based on the current format. If `formats` is empty or not provided, the clock uses the single `format` string as before.